Move tests under spring.composer to align with package structure
