+++ BUILD.bazel
@@ -208,9 +208,9 @@ GOLDEN_UPDATING_UNIT_TESTS = [
     "com.google.api.generator.gapic.composer.rest.ServiceClientTestClassComposerTest",
     "com.google.api.generator.gapic.composer.rest.ServiceSettingsClassComposerTest",
     "com.google.api.generator.gapic.composer.rest.ServiceStubSettingsClassComposerTest",
-    "com.google.api.generator.spring.SpringAutoConfigClassComposerTest",
-    "com.google.api.generator.spring.SpringPropertiesClassComposerTest",
-    "com.google.api.generator.spring.SpringComposerTest",
+    "com.google.api.generator.spring.composer.SpringAutoConfigClassComposerTest",
+    "com.google.api.generator.spring.composer.SpringPropertiesClassComposerTest",
+    "com.google.api.generator.spring.composer.SpringComposerTest",
 ]
 
 # Manually runs a single JUnit test for the tests that generate and save .golden
