+++ src/test/java/com/google/api/generator/spring/composer/goldens/EchoSpringPropertiesFull.golden
